+++ pkg/sql/colexec/any_not_null_agg_tmpl.go
@@ -105,7 +105,7 @@ func (a *anyNotNull_TYPEAgg) CurrentOutputIndex() int {
 func (a *anyNotNull_TYPEAgg) SetOutputIndex(idx int) {
 	if a.curIdx != -1 {
 		a.curIdx = idx
-		a.nulls.UnsetNullsAfter(uint16(idx + 1))
+		a.nulls.UnsetNullsAfter(uint64(idx + 1))
 	}
 }
 
@@ -118,7 +118,7 @@ func (a *anyNotNull_TYPEAgg) Compute(b coldata.Batch, inputIdxs []uint32) {
 		// If we haven't found any non-nulls for this group so far, the output for
 		// this group should be null.
 		if !a.foundNonNullForCurrentGroup {
-			a.nulls.SetNull(uint16(a.curIdx))
+			a.nulls.SetNull(uint64(a.curIdx))
 		} else {
 			execgen.SET(a.col, a.curIdx, a.curAgg)
 		}
@@ -182,7 +182,7 @@ func _FIND_ANY_NOT_NULL(a *anyNotNull_TYPEAgg, nulls *coldata.Nulls, i int, _HAS
 			// If this is a new group, check if any non-nulls have been found for the
 			// current group.
 			if !a.foundNonNullForCurrentGroup {
-				a.nulls.SetNull(uint16(a.curIdx))
+				a.nulls.SetNull(uint64(a.curIdx))
 			} else {
 				execgen.SET(a.col, a.curIdx, a.curAgg)
 			}
@@ -192,7 +192,7 @@ func _FIND_ANY_NOT_NULL(a *anyNotNull_TYPEAgg, nulls *coldata.Nulls, i int, _HAS
 	}
 	var isNull bool
 	// {{ if .HasNulls }}
-	isNull = nulls.NullAt(uint16(i))
+	isNull = nulls.NullAt(uint64(i))
 	// {{ else }}
 	isNull = false
 	// {{ end }}
